Move CommentEntry to Comment
