+++ src/ui/components/Comments/Comment.js
@@ -1,135 +1,116 @@
-import React from "react";
+import React, { useEffect, useState, useRef } from "react";
 import { connect } from "react-redux";
 import classnames from "classnames";
-
 import { selectors } from "ui/reducers";
 import { actions } from "ui/actions";
-import { getPixelOffset, getCommentLeftOffset } from "ui/utils/timeline";
 
-import CommentMarker from "./CommentMarker";
+import CommentEditor from "ui/components/Comments/CommentEditor";
+import CommentDropdownPanel from "ui/components/Comments/CommentDropdownPanel";
 import PortalDropdown from "ui/components/shared/PortalDropdown";
-import CommentEditor from "./CommentEditor";
-import CommentDropdownPanel from "./CommentDropdownPanel";
 
-class Comment extends React.Component {
-  state = {
-    editing: false,
+function Comment({ comment, user, currentTime, seek }) {
+  const [editing, setEditing] = useState(false);
+  const [replying, setReplying] = useState(false);
+  const commentEl = useRef(null);
+  const [menuExpanded, setMenuExpanded] = useState(false);
+  const seekToComment = () => {
+    const { point, time, has_frames } = comment;
+
+    if (editing) {
+      return;
+    }
+
+    return seek(point, time, has_frames);
   };
 
-  componentDidMount() {
-    const { comment } = this.props;
-
+  useEffect(() => {
     // A newly-added comment, which is initialized as empty, should go directly
     // into editing mode.
-    if (!comment.contents) {
-      this.startEditing();
+    if (!comment.content) {
+      setEditing(true);
     }
-  }
+  }, [comment]);
 
-  componentDidUpdate() {
-    const { comment, focusedCommentId } = this.props;
-    const notFocused = focusedCommentId !== comment.id;
-
-    // We make sure that all unfocused comments are not in their editing state.
-    // This way when a user focuses on a comment, the editor is not displayed.
-    if (notFocused && this.state.editing) {
-      this.stopEditing();
+  useEffect(() => {
+    if (comment.time === currentTime) {
+      commentEl.current.scrollIntoView({ behavior: "smooth", block: "center" });
     }
-  }
+  }, [currentTime]);
 
-  startEditing = () => {
-    const { user, comment } = this.props;
+  return (
+    <div className="comment-container" ref={commentEl}>
+      <div
+        className={classnames(
+          "comment",
+          { selected: currentTime === comment.time },
+          { reply: replying },
+          { "child-comment": comment.parent_id }
+        )}
+        onClick={seekToComment}
+      >
+        <img src={comment.user.picture} className="comment-picture" />
+        {editing ? (
+          <CommentEditor replying={false} comment={comment} stopEditing={() => setEditing(false)} />
+        ) : (
+          <CommentBody comment={comment} user={user} startEditing={() => setEditing(true)} />
+        )}
+        <div className="comment-dropdown" onClick={e => e.stopPropagation()}>
+          <PortalDropdown
+            buttonContent={<div className="dropdown-button">⋯</div>}
+            expanded={menuExpanded}
+            setExpanded={setMenuExpanded}
+          >
+            <CommentDropdownPanel
+              user={user}
+              comment={comment}
+              allowReply={true}
+              startEditing={() => setEditing(true)}
+              startReplying={() => setReplying(true)}
+              onItemClick={() => setMenuExpanded(false)}
+            />
+          </PortalDropdown>
+        </div>
+      </div>
+      {replying && (
+        <div className="comment">
+          <CommentEditor
+            replying={true}
+            comment={comment}
+            stopReplying={() => setReplying(false)}
+          />
+        </div>
+      )}
+    </div>
+  );
+}
+
+function CommentBody({ comment, user, startEditing }) {
+  const lines = comment.content.split("\n");
+
+  const onDoubleClick = () => {
     if (user?.loggedIn && comment.user_id == user?.id) {
-      this.setState({ editing: true });
+      startEditing();
     }
   };
 
-  stopEditing = () => {
-    this.setState({ editing: false });
-  };
-
-  deleteComment = () => {
-    const { setFocusedCommentId, comment } = this.props;
-    setFocusedCommentId(null);
-    removeComment(comment);
-  };
-
-  renderDropdownPanel() {
-    const { comment, user } = this.props;
-    return (
-      <CommentDropdownPanel
-        startEditing={this.startEditing}
-        allowReply={false}
-        user={user}
-        comment={comment}
-        onItemClick={() => this.setState({ menuExpanded: false })}
-      />
-    );
-  }
-
-  renderLabel() {
-    const { comment } = this.props;
-    const lines = comment.content.split("\n");
-
-    return (
-      <div className="label" onDoubleClick={this.startEditing}>
+  return (
+    <div className="comment-body">
+      <div className="item-label">{comment.user.name}</div>
+      <div className="item-content" onDoubleClick={onDoubleClick}>
         {lines.map((line, i) => (
           <div key={i}>{line}</div>
         ))}
       </div>
-    );
-  }
-
-  renderCommentBody() {
-    const { editing, menuExpanded } = this.state;
-    const { comment } = this.props;
-    const isNewComment = comment.content === "";
-
-    return (
-      <div className="comment-body">
-        <div className="comment-content">
-          {!isNewComment ? (
-            <div className="comment-header">
-              <div className="actions">
-                <PortalDropdown
-                  buttonContent={<div className="dropdown-button">⋯</div>}
-                  expanded={menuExpanded}
-                  setExpanded={value => this.setState({ menuExpanded: value })}
-                >
-                  {this.renderDropdownPanel()}
-                </PortalDropdown>
-              </div>
-            </div>
-          ) : null}
-          <div className="comment-description">
-            {editing ? (
-              <CommentEditor comment={comment} stopEditing={this.stopEditing} />
-            ) : (
-              this.renderLabel()
-            )}
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  render() {
-    const { comment } = this.props;
-
-    return <CommentMarker comment={comment} />;
-  }
+    </div>
+  );
 }
 
 export default connect(
   state => ({
-    timelineDimensions: selectors.getTimelineDimensions(state),
-    zoomRegion: selectors.getZoomRegion(state),
     currentTime: selectors.getCurrentTime(state),
-    focusedCommentId: selectors.getFocusedCommentId(state),
     user: selectors.getUser(state),
   }),
   {
-    removeComment: actions.removeComment,
-    setFocusedCommentId: actions.setFocusedCommentId,
+    seek: actions.seek,
   }
 )(Comment);
