+++ src/ui/components/SecondaryToolbox/TranscriptEntry/TranscriptEntry.js
@@ -1,10 +1,10 @@
 import React from "react";
-import CommentEntry from "./CommentEntry";
+import Comment from "ui/components/Comments/Comment";
 import EventEntry from "./EventEntry";
 
 export default function TranscriptEntry({ entry }) {
   if (entry.__typename === "comments") {
-    return <CommentEntry comment={entry} />;
+    return <Comment comment={entry} />;
   } else {
     return <EventEntry entry={entry} />;
   }
